+++ server/src/main/java/luan/moonvs/models/requests/UserAccountRequest.java
@@ -1,8 +1,13 @@
 package luan.moonvs.models.requests;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+import java.time.LocalDate;
 import java.util.UUID;
 
+@JsonIgnoreProperties
 public record UserAccountRequest(UUID idUser,
                                  String username,
                                  String email,
+                                 LocalDate birthDate,
                                  String password) { }